refactor: part 2 of /memberships

+++ src/server/groups/memberships/destroy.ts
@@ -2,12 +2,16 @@ import 'server-only'
 import { prismaCall } from '@/server/prismaCall'
 import type { ExpandedMembership } from './Types'
 import { readCurrenOmegaOrder } from '@/server/omegaOrder/read'
+import { canEasalyManageMembershipOfGroup } from './canEasalyManageMembership'
+import { ServerError } from '@/server/error'
 
 export async function destoryMembershipOfUser(
     groupId: number,
     userId: number,
     orderArg?: number
 ): Promise<ExpandedMembership> {
+    if (!canEasalyManageMembershipOfGroup(groupId))
+        throw new ServerError('BAD PARAMETERS', 'Denne Gruppetypen kan ikke enkelt opprette medlemskap')
     const order = orderArg ?? (await readCurrenOmegaOrder()).order
 
     return await prismaCall(() => prisma.membership.delete({
@@ -26,6 +30,8 @@ export async function destroyMembershipOfUsers(
     userIds: number[],
     orderArg?: number,
 ): Promise<void> {
+    if (!canEasalyManageMembershipOfGroup(groupId))
+        throw new ServerError('BAD PARAMETERS', 'Denne Gruppetypen kan ikke enkelt opprette medlemskap')
     const order = orderArg ?? (await readCurrenOmegaOrder()).order
 
     await prismaCall(() => prisma.membership.deleteMany({
